850: delete sample module - Sample module deleted - Moved docker build to server module
Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/850
