+++ pom.xml
@@ -31,7 +31,6 @@
     <modules>
         <module>secure-api-gateway-ob-uk-rcs-api</module>
         <module>secure-api-gateway-ob-uk-rcs-server</module>
-        <module>secure-api-gateway-ob-uk-rcs-sample</module>
         <module>secure-api-gateway-ob-uk-rcs-cloud-client</module>
     </modules>
