--- secure-api-gateway-ob-uk-rcs-sample/README.md
@@ -1,10 +1,0 @@
-## secure-api-gateway-ob-uk-rcs-sample
-
-A Sample Spring Boot application to start up the UK Open Banking Remote Consent Service for the Secure API Gateway.
-
-## check the spring cloud config server properties
-
-> profile docker
-
-- http://localhost:8888/securebanking-openbanking-uk-rcs/docker
-- http://localhost:8888/securebanking-openbanking-uk-rs/docker
