+++ secure-api-gateway-ob-uk-rcs-server/docker/scripts/wait-for.sh
