+++ secure-api-gateway-ob-uk-rcs-server/src/main/resources/application.yml
